fix(ci+transport): unbreak the CI matrix after runner-image drift
Four targeted fixes to get PR #1 (docs-only) past CI. None of these
were caused by the docs changes — main's CI was green yesterday and
the underlying source is unchanged, so the runner-image roll-forward
on the Ubuntu / macOS / Windows hosted runners tightened toolchain
checks that exposed dormant issues.

1. .github/workflows/ci.yml: install wayland-protocols + libwayland-dev
   on both linux-asan and linux-gcc jobs. GLFW's CMake config now
   looks for wayland-scanner during configure on Ubuntu 24.04, and
   the workflow wasn't installing the Wayland dev packages. The fix
   matches what Ubuntu 24.04's image used to ship pre-installed.

2. src/transport/src/obdx_transport.cpp:371-372 — MSVC rejects the
   implicit conversion from std::array<uint8_t, 0> to std::span<uint8_t
   const>. Switch to constructing an empty std::span directly. GCC /
   Clang accepted the array form; MSVC's stricter span constructor
   rules don't. Behaviour-identical; the TODO context above the call
   already explains the payload is intentionally empty pending PDF
   verification.

3. src/transport/src/native.cpp:27 — Apple Clang's sign-conversion
   warning fires on (crc << 1U) ^ 0x1021U because the shift result is
   int (integer promotion) and the XOR with unsigned int produces a
   signed/unsigned mismatch before the outer uint16_t cast. Wrap the
   shift in static_cast<std::uint16_t> first so the XOR is unsigned
   on both sides. Same numerical result for all inputs.

4. src/transport/src/native.cpp:108 — Same family of issue on the CRC
   extraction. Wrap the whole shift-or expression in static_cast<
   std::uint16_t> so the narrowing on assignment is explicit.

None of these change behaviour; they make the existing intent compile
under tighter toolchains. 828 C++ unit-test cases stay green locally.

Note: the clang-format (required) job is also failing on this PR but
with 195 pre-existing violations spanning 10 files (src/cli/main.cpp,
src/flash/, src/transport/uds_trace.*, src/ui/src/main.cpp, and 3 test
files). That's a separate fix — either a one-shot mass reformat against
clang-format 18.1.8 or pinning the CI clang-format version via pip
instead of apt. Not in scope for this PR; tracked for a follow-up.

+++ src/transport/src/native.cpp
@@ -24,7 +24,7 @@ std::uint16_t crc16_ccitt_false(std::span<std::uint8_t const> bytes) noexcept {
         crc ^= static_cast<std::uint16_t>(b) << 8U;
         for (int i = 0; i < 8; ++i) {
             if ((crc & 0x8000U) != 0) {
-                crc = static_cast<std::uint16_t>((crc << 1U) ^ 0x1021U);
+                crc = static_cast<std::uint16_t>(static_cast<std::uint16_t>(crc << 1U) ^ 0x1021U);
             } else {
                 crc = static_cast<std::uint16_t>(crc << 1U);
             }
@@ -105,8 +105,9 @@ Result<DecodedFrame> decode_frame(std::span<std::uint8_t const> bytes) {
     }
 
     std::uint16_t const expected_crc = crc_over_frame(bytes.subspan(0, bytes.size() - 2));
-    std::uint16_t const actual_crc = (static_cast<std::uint16_t>(bytes[bytes.size() - 2]) << 8U) |
-                                     static_cast<std::uint16_t>(bytes[bytes.size() - 1]);
+    std::uint16_t const actual_crc =
+        static_cast<std::uint16_t>((static_cast<std::uint16_t>(bytes[bytes.size() - 2]) << 8U) |
+                                   static_cast<std::uint16_t>(bytes[bytes.size() - 1]));
     if (expected_crc != actual_crc) {
         char buf[96];
         std::snprintf(buf, sizeof buf,
